requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiohttp-cors==0.8.1
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appnope==0.1.4
apsw==3.51.3.0
apswutils==0.1.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
autogluon-common==1.5.0
autogluon-core==1.5.0
autogluon-features==1.5.0
autogluon-tabular==1.5.0
babel==2.18.0
beartype==0.22.9
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
blis==1.3.3
boto3==1.42.78
botocore==1.42.78
cachetools==7.0.5
catalogue==2.0.10
catboost==1.2.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.2.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorful==0.5.8
comm==0.2.3
confection==1.3.3
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
einops==0.8.2
einx==0.4.2
executing==2.2.1
fastai==2.8.7
fastcore==1.12.31
fastdownload==0.0.7
fastjsonschema==2.21.2
fastlite==0.2.4
fastprogress==1.1.5
fasttransform==0.0.2
filelock==3.25.2
fonttools==4.62.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.1
graphviz==0.21
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.2
hyperopt==0.2.7
idna==3.11
importlib-metadata==8.7.1
ipykernel==7.2.0
ipython==9.12.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.8.0
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lightgbm==4.6.0
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.15
narwhals==2.18.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook-shim==0.2.4
numba==0.64.0
numpy==2.3.5
oauthlib==3.3.1
omegaconf==2.3.0
openai==2.30.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.40.0
opentelemetry-exporter-prometheus==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pbr==7.0.3
pexpect==4.9.0
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.4
plotly==6.6.0
plum-dispatch==2.8.0
preshed==3.0.13
prometheus-client==0.24.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
py-spy==0.4.1
py4j==0.10.9.9
pyarrow==20.0.0
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-dotenv==1.2.2
python-fasthtml==0.13.2
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2026.1.post1
pyyaml==6.0.3
pyzmq==27.1.0
ray==2.52.1
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
send2trash==2.1.0
setuptools==82.0.1
shap==0.51.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart-open==7.5.1
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
stack-data==0.6.3
starlette==1.0.0
stevedore==5.4.1
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
tensorboardx==2.6.4
terminado==0.18.1
thinc==8.3.13
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
virtualenv==21.2.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==1.0.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wheel==0.46.3
widgetsnbextension==4.0.15
wrapt==2.1.2
xgboost==3.1.3
yarl==1.23.0
zipp==3.23.0
